Added missing flags to command-line parser. Fixed some virtual classes without virtual destructor. Fixed playback S/E points in EDL playback.  Note that playback in reverse is still broken.
Former-commit-id: 744a118989c12c173a6e545d1b195dc593e45bfc [formerly 84b9b6d9aaaa054cbf459b8ee8d755206a38c7fc [formerly 3a64dd5fc40b58eb05295f6c6a0178b5e4b53883]]
Former-commit-id: 0ef60dcbe6b56708273a15e515ae0ab730007b04 [formerly 7302488bd165c4bfb39b29076e83d8629189e2af] [formerly 2ee4d155f3618bf6f3fa0e6a3d33fac17e1d7150 [formerly 16900b218fd2ff6fb7b4212f11dd6976d1cfe108]]
Former-commit-id: a3a73576de391d666f969e843ade734218122ce5 [formerly 5e7364c4b37dcd0f3e0a47b1ca33541bfe23d998]
Former-commit-id: b77f116ffb8b6fbb6f4eed9e27a89d56ed01f142

+++ mrViewer/src/core/guessImage.cpp
@@ -126,7 +126,7 @@ CMedia* test_image( const char* name,
 
 std::string parse_view( const std::string& root, bool left )
 {
-    int idx = root.find( "%V" );
+    size_t idx = root.find( "%V" );
     std::string tmp = root;
     if ( idx != std::string::npos )
     {
